+++ src/Vector/ASC/CanError.cpp
@@ -31,7 +31,9 @@ CanError::CanError() :
     Event(),
     time(0.0),
     channel(0),
-    error()
+    error(),
+    txErr(0),
+    rxErr(0)
 {
     eventType = EventType::CanError;
 }
@@ -44,7 +46,41 @@ CanError * CanError::read(File & /*file*/, std::string & line)
         CanError * canError = new CanError;
         canError->time = std::stod(match[1]);
         canError->channel = std::stoul(match[2]);
-        canError->error = match[3];
+        std::string errorStr = match[3];
+        size_t rxErrPos = errorStr.find(" RxErr: ");
+        if (rxErrPos != std::string::npos) {
+            canError->rxErr = std::stoul(errorStr.substr(rxErrPos + 8));
+            errorStr.erase(rxErrPos);
+        }
+        size_t txErrPos = errorStr.find(" TxErr: ");
+        if (txErrPos != std::string::npos) {
+            canError->txErr = std::stoul(errorStr.substr(txErrPos + 8));
+            errorStr.erase(txErrPos);
+        }
+        size_t minusPos = errorStr.find(" -");
+        if (minusPos != std::string::npos) {
+            errorStr.erase(minusPos);
+        }
+        if (errorStr == "chip status busoff") {
+            canError->error = Error::ChipStatusBusoff;
+        } else
+        if (errorStr == "chip status error passive") {
+            canError->error = Error::ChipStatusErrorPassive;
+        } else
+        if (errorStr == "chip status error active") {
+            canError->error = Error::ChipStatusErrorActive;
+        } else
+        if (errorStr == "rx queue overrun") {
+            canError->error = Error::RxQueueOverrun;
+        } else
+        if (errorStr == "chip status warning level") {
+            canError->error = Error::ChipStatusWarningLevel;
+        } else
+        if (errorStr == "hardware interface disconnected") {
+            canError->error = Error::HardwareInterfaceDisconnected;
+        } else {
+            canError->error = Error::Unknown;
+        }
         return canError;
     }
 
@@ -61,19 +97,48 @@ void CanError::write(File & file, std::ostream & stream)
                 << " CAN " << std::dec << (int16_t) channel
                 << " Status: ";
     } else {
-        /* "format: "CAN %d Status: " */
+        /* "format: "CAN %d Status:" */
         stream
                 << " CAN " << std::dec << (int16_t) channel
                 << " Status:";
     }
 
-    /* format: "chip status busoff" */
-    /* format: "chip status error passive" */
-    /* format: "chip status error active" */
-    /* format: "rx queue overrun" */
-    /* format: "chip status warning level" */
-    /* format: "hardware interface disconnected" */
-    stream << error;
+    switch(error) {
+    case Error::Unknown:
+        break;
+    case Error::ChipStatusBusoff:
+        /* format: "chip status busoff" */
+        stream << "chip status busoff";
+        break;
+    case Error::ChipStatusErrorPassive:
+        /* format: "chip status error passive" */
+        stream << "chip status error passive";
+        break;
+    case Error::ChipStatusErrorActive:
+        /* format: "chip status error active" */
+        stream << "chip status error active";
+        break;
+    case Error::RxQueueOverrun:
+        /* format: "rx queue overrun" */
+        stream << "rx queue overrun";
+        break;
+    case Error::ChipStatusWarningLevel:
+        /* format: "chip status warning level" */
+        stream << "chip status warning level";
+        break;
+    case Error::HardwareInterfaceDisconnected:
+        /* format: "hardware interface disconnected" */
+        stream << "hardware interface disconnected";
+        break;
+    }
+
+    /* format: "%s - TxErr: %d RxErr: %d" */
+    if ((txErr > 0) || (rxErr > 0)) {
+        stream
+                << " -"
+                << " TxErr: " << std::dec << (uint16_t) txErr
+                << " RxErr: " << std::dec << (uint16_t) rxErr;
+    }
 
     stream << endl;
 }